Subtraction operator and part of multiplication operator
The lexer was changed to allow negative numbers.

The subtraction (`-`) operator was implemented.

Part of the multiplication (`*`) operator was implemented. Block
execution needs to be implemented before multiplication can be
finished. The problem with this is that binary_operator needs an object
to push to the stack, which executing won't return. I've thought about
adding a `None` to the object type, but that would make things a lot
more complicated.

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>